refactor: Split unit tests into separate files
