+++ tests/base.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2020 Famedly
+# Copyright (C) 2020, 2024 Famedly
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published by
@@ -18,7 +18,7 @@ import json
 from typing import Any
 from unittest.mock import AsyncMock, Mock, patch
 
-from synapse.module_api import NOT_SPAM, errors
+from synapse.module_api import errors
 from synapse.rest import admin
 from synapse.rest.client import (
     account_data,
@@ -30,13 +30,11 @@ from synapse.rest.client import (
 )
 from synapse.server import HomeServer
 from synapse.util import Clock
-from twisted.internet import defer
 from twisted.internet.testing import MemoryReactor
 from typing_extensions import override
 
 import tests.unittest as synapsetest
-from synapse_invite_checker import InviteChecker
-from synapse_invite_checker.types import FederationDomain
+from tests.test_utils import DOMAIN_IN_LIST
 
 # ruff: noqa: E501
 # We don't care about long lines in our testdata
@@ -125,10 +123,6 @@ def return_gem_cert(cn: str) -> bytes:
     if cn == "GEM.KOMP-CA50 TEST-ONLY":
         return base64.b64decode(gem_komp_ca50_cert)
     raise Exception("Could not find cert" + cn)
-
-
-# Use a domain found in our raw fedlist data instead of inserting a fake one
-DOMAIN_IN_LIST = "timo.staging.famedly.de"
 
 
 def return_localization(mxid: str) -> str:
@@ -250,723 +244,3 @@ class ModuleApiTestCase(synapsetest.HomeserverTestCase):
                 }
             ]
         return conf
-
-
-class LocalInviteTest(ModuleApiTestCase):
-    def prepare(self, reactor: MemoryReactor, clock: Clock, homeserver: HomeServer):
-        super().prepare(reactor, clock, homeserver)
-        self.user_a = self.register_user("a", "password")
-        self.access_token = self.login("a", "password")
-        self.user_b = self.register_user("b", "password")
-        self.user_c = self.register_user("c", "password")
-
-        # authenticated as user_a
-        self.helper.auth_user_id = self.user_a
-
-    def test_invite_to_dm(self) -> None:
-        """Tests that a dm with a local user can be created, but nobody else invited"""
-        room_id = self.helper.create_room_as(
-            self.user_a, is_public=False, tok=self.access_token
-        )
-        assert room_id, "Room created"
-
-        # create DM event
-        channel = self.make_request(
-            "PUT",
-            f"/user/{self.user_a}/account_data/m.direct",
-            {
-                self.user_b: [room_id],
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-
-        # Can't invite other users
-        self.helper.invite(
-            room=room_id,
-            src=self.user_a,
-            targ=self.user_c,
-            tok=self.access_token,
-            expect_code=403,
-        )
-        # But can invite the dm user
-        self.helper.invite(
-            room=room_id,
-            src=self.user_a,
-            targ=self.user_b,
-            tok=self.access_token,
-            expect_code=200,
-        )
-
-    def test_invite_to_group(self) -> None:
-        """Tests that a group with local users works normally"""
-        room_id = self.helper.create_room_as(
-            self.user_a, is_public=False, tok=self.access_token
-        )
-        assert room_id, "Room created"
-
-        # create DM event
-        channel = self.make_request(
-            "PUT",
-            f"/user/{self.user_a}/account_data/m.direct",
-            {
-                self.user_b: ["!not:existing.example.com"],
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-
-        # Can invite other users
-        self.helper.invite(
-            room=room_id,
-            src=self.user_a,
-            targ=self.user_c,
-            tok=self.access_token,
-            expect_code=200,
-        )
-        self.helper.invite(
-            room=room_id,
-            src=self.user_a,
-            targ=self.user_b,
-            tok=self.access_token,
-            expect_code=200,
-        )
-
-    def test_invite_to_group_without_dm_event(self) -> None:
-        """Tests that a group with local users works normally in case the user has no m.direct set"""
-        room_id = self.helper.create_room_as(
-            self.user_a, is_public=False, tok=self.access_token
-        )
-        assert room_id, "Room created"
-
-        # Can invite other users
-        self.helper.invite(
-            room=room_id,
-            src=self.user_a,
-            targ=self.user_c,
-            tok=self.access_token,
-            expect_code=200,
-        )
-        self.helper.invite(
-            room=room_id,
-            src=self.user_a,
-            targ=self.user_b,
-            tok=self.access_token,
-            expect_code=200,
-        )
-
-
-class RemoteInviteTest(ModuleApiTestCase):
-    def prepare(self, reactor: MemoryReactor, clock: Clock, homeserver: HomeServer):
-        super().prepare(reactor, clock, homeserver)
-        self.user_a = self.register_user("a", "password")
-        self.access_token = self.login("a", "password")
-        self.user_b = self.register_user("b", "password")
-        self.user_c = self.register_user("c", "password")
-
-        # authenticated as user_a
-        self.helper.auth_user_id = self.user_a
-
-    async def may_invite(self, inviter: str, invitee: str, roomid: str):
-        req = defer.ensureDeferred(
-            self.hs.get_module_api()._callbacks.spam_checker.user_may_invite(
-                inviter, invitee, roomid
-            )
-        )
-        self.wait_on_thread(req)
-        ret = self.get_success(req)
-        if ret == NOT_SPAM:
-            return NOT_SPAM
-        return ret[0]  # return first code instead of all of them to make assert easier
-
-    async def test_invite_from_remote_outside_of_fed_list(self) -> None:
-        """Tests that an invite from a remote server not in the federation list gets denied"""
-
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": f"@example:{DOMAIN_IN_LIST}",
-                "inviteSettings": {
-                    "start": 0,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": f"@example:not-{DOMAIN_IN_LIST}",
-                "inviteSettings": {
-                    "start": 0,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-
-        assert (
-            await self.may_invite(
-                f"@example:not-{DOMAIN_IN_LIST}", self.user_a, "!madeup:example.com"
-            )
-            == errors.Codes.FORBIDDEN
-        )
-        # currently not testing modifying the fed list atm
-        assert (
-            await self.may_invite(
-                "@example:messenger.spilikin.dev", self.user_a, "!madeup:example.com"
-            )
-            == errors.Codes.FORBIDDEN
-        )
-        assert (
-            await self.may_invite(
-                f"@example:{DOMAIN_IN_LIST}", self.user_a, "!madeup:example.com"
-            )
-            == NOT_SPAM
-        )
-        assert (
-            await self.may_invite(
-                f"@example2:not-{DOMAIN_IN_LIST}", self.user_a, "!madeup:example.com"
-            )
-            == errors.Codes.FORBIDDEN
-        )
-        assert (
-            await self.may_invite(
-                f"@example2:{DOMAIN_IN_LIST}", self.user_a, "!madeup:example.com"
-            )
-            == errors.Codes.FORBIDDEN
-        )
-
-    async def test_invite_from_publicly_listed_practitioners(self) -> None:
-        """Tests that an invite from a remote server gets accepted when in the federation list and both practitioners are public"""
-        for inviter in {
-            f"@mxid:{DOMAIN_IN_LIST}",
-            f"@matrixuri:{DOMAIN_IN_LIST}",
-            f"@matrixuri2:{DOMAIN_IN_LIST}",
-            f"@gematikuri:{DOMAIN_IN_LIST}",
-            f"@gematikuri2:{DOMAIN_IN_LIST}",
-            f"@mxidorgpract:{DOMAIN_IN_LIST}",
-            f"@matrixuriorgpract:{DOMAIN_IN_LIST}",
-            f"@matrixuri2orgpract:{DOMAIN_IN_LIST}",
-            f"@gematikuriorgpract:{DOMAIN_IN_LIST}",
-            f"@gematikuri2orgpract:{DOMAIN_IN_LIST}",
-        }:
-            assert (
-                await self.may_invite(inviter, self.user_a, "!madeup:example.com")
-                == NOT_SPAM
-            )
-            assert (
-                await self.may_invite(inviter, self.user_c, "!madeup:example.com")
-                == NOT_SPAM
-            )
-
-        for inviter in {
-            f"@mxid404:{DOMAIN_IN_LIST}",
-            f"@matrixuri404:{DOMAIN_IN_LIST}",
-            f"@matrixuri2404:{DOMAIN_IN_LIST}",
-            f"@gematikuri404:{DOMAIN_IN_LIST}",
-            f"@gematikuri2404:{DOMAIN_IN_LIST}",
-        }:
-            assert (
-                await self.may_invite(inviter, self.user_a, "!madeup:example.com")
-                == errors.Codes.FORBIDDEN
-            )
-
-    async def test_invite_from_remote_to_local_org(self) -> None:
-        """Tests that an invite from a remote server gets accepted when in the federation list and the invite is to an orgPract"""
-        for inviter in {
-            f"@mxid:{DOMAIN_IN_LIST}",
-            f"@matrixuri:{DOMAIN_IN_LIST}",
-            f"@matrixuri2:{DOMAIN_IN_LIST}",
-            f"@gematikuri:{DOMAIN_IN_LIST}",
-            f"@gematikuri2:{DOMAIN_IN_LIST}",
-            f"@mxidorgpract:{DOMAIN_IN_LIST}",
-            f"@matrixuriorgpract:{DOMAIN_IN_LIST}",
-            f"@matrixuri2orgpract:{DOMAIN_IN_LIST}",
-            f"@gematikuriorgpract:{DOMAIN_IN_LIST}",
-            f"@gematikuri2orgpract:{DOMAIN_IN_LIST}",
-            f"@mxid404:{DOMAIN_IN_LIST}",
-            f"@matrixuri404:{DOMAIN_IN_LIST}",
-            f"@matrixuri2404:{DOMAIN_IN_LIST}",
-            f"@gematikuri404:{DOMAIN_IN_LIST}",
-            f"@gematikuri2404:{DOMAIN_IN_LIST}",
-        }:
-            assert (
-                await self.may_invite(inviter, self.user_b, "!madeup:example.com")
-                == NOT_SPAM
-            )
-            assert (
-                await self.may_invite(inviter, self.user_c, "!madeup:example.com")
-                == NOT_SPAM
-            )
-
-        assert (
-            await self.may_invite(
-                "@unknown:not.in.fed", self.user_b, "!madeup:example.com"
-            )
-            == errors.Codes.FORBIDDEN
-        )
-        assert (
-            await self.may_invite(
-                "@unknown:not.in.fed", self.user_c, "!madeup:example.com"
-            )
-            == errors.Codes.FORBIDDEN
-        )
-
-
-class ContactsApiTest(ModuleApiTestCase):
-    def prepare(self, reactor: MemoryReactor, clock: Clock, homeserver: HomeServer):
-        super().prepare(reactor, clock, homeserver)
-        self.user_a = self.register_user("a", "password")
-        self.access_token = self.login("a", "password")
-        self.user_b = self.register_user("b", "password")
-        self.user_c = self.register_user("c", "password")
-
-        # authenticated as user_a
-        self.helper.auth_user_id = self.user_a
-
-    async def test_contact_management_info_resource(self) -> None:
-        """Tests that the registered info resource is accessible"""
-
-        channel = self.make_request(
-            method="GET",
-            path="/_synapse/client/com.famedly/tim/v1",
-        )
-
-        assert channel.code == 200, channel.result
-        assert channel.json_body["title"] == "Invite Checker module by Famedly"
-        assert channel.json_body["description"] == "Invite Checker module by Famedly"
-        assert channel.json_body["contact"] == "info@famedly.com"
-        assert channel.json_body["version"], "Version returned"
-
-    def test_list_contacts(self) -> None:
-        # Requires authentication
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-        )
-        assert channel.code == 401, channel.result
-
-        # Test for a reasonable return value
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "contacts" in channel.json_body, "Result contains contacts"
-        assert len(channel.json_body["contacts"]) == 0, "List is empty initially"
-
-    def test_add_contact(self) -> None:
-        """See if we can add a new contact"""
-        # Requires authentication
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": "@test:other.example.com",
-                "inviteSettings": {"start": 0},
-            },
-        )
-        assert channel.code == 401, channel.result
-
-        # Test missing field validation
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "inviteSettings": {
-                    "start": 0,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 400, channel.result
-
-        # Test for a reasonable return value
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": "@test:other.example.com",
-                "inviteSettings": {
-                    "start": 0,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "displayName" in channel.json_body, "Result contains contact"
-
-        # check for 1 contact
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "contacts" in channel.json_body, "Result contains contacts"
-        assert len(channel.json_body["contacts"]) == 1, "List is not empty anymore"
-
-        # see if we properly overwrite settings
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": "@test:other.example.com",
-                "inviteSettings": {
-                    "start": 500,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "displayName" in channel.json_body, "Result contains contact"
-
-        # check for 1 contact
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "contacts" in channel.json_body, "Result contains contacts"
-        assert len(channel.json_body["contacts"]) == 1, "List is not empty anymore"
-        assert (
-            channel.json_body["contacts"][0]["inviteSettings"]["start"] == 500
-        ), "Setting overwritten"
-
-        # And post works as well
-        channel = self.make_request(
-            "POST",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": "@test:other.example.com",
-                "inviteSettings": {
-                    "start": 400,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "displayName" in channel.json_body, "Result contains contact"
-
-        # check for 1 contact
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "contacts" in channel.json_body, "Result contains contacts"
-        assert len(channel.json_body["contacts"]) == 1, "List is not empty anymore"
-        assert (
-            channel.json_body["contacts"][0]["inviteSettings"]["start"] == 400
-        ), "Setting overwritten"
-
-        # And we can add a second user
-        channel = self.make_request(
-            "POST",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User2",
-                "mxid": "@test2:other.example.com",
-                "inviteSettings": {
-                    "start": 300,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "displayName" in channel.json_body, "Result contains contact"
-
-        # check for 2 contacts
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "contacts" in channel.json_body, "Result contains contacts"
-        assert len(channel.json_body["contacts"]) == 2, "List is not empty anymore"
-        assert [
-            item["inviteSettings"]["start"]
-            for item in channel.json_body["contacts"]
-            if item["mxid"] == "@test:other.example.com"
-        ] == [400], "Setting overwritten"
-        assert [
-            item["inviteSettings"]["start"]
-            for item in channel.json_body["contacts"]
-            if item["mxid"] == "@test2:other.example.com"
-        ] == [300], "Setting overwritten"
-
-    def test_get_contact(self) -> None:
-        """See if we can retrieve a specific contact"""
-        # Requires authentication
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test:other.example.com",
-        )
-        assert channel.code == 401, channel.result
-
-        # 404 on unknown contact
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test:other.example.com",
-            access_token=self.access_token,
-        )
-        assert channel.code == 404, channel.result
-
-        # Test for a reasonable return value
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": "@test:other.example.com",
-                "inviteSettings": {
-                    "start": 0,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "displayName" in channel.json_body, "Result contains contact"
-
-        # 200 after adding
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test:other.example.com",
-            access_token=self.access_token,
-        )
-
-        # Verify that adding another still returns a value
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": "@test2:other.example.com",
-                "inviteSettings": {
-                    "start": 0,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "displayName" in channel.json_body, "Result contains contact"
-
-        # 200 after adding
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test:other.example.com",
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-
-        # Unknown user still returns 404 after adding
-        channel = self.make_request(
-            "GET",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test3:other.example.com",
-            access_token=self.access_token,
-        )
-        assert channel.code == 404, channel.result
-
-    def test_del_contact(self) -> None:
-        """See if we can delete a contact"""
-        # Requires authentication
-        channel = self.make_request(
-            "DELETE",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test:other.example.com",
-        )
-        assert channel.code == 401, channel.result
-
-        # 404 on unknown contact
-        channel = self.make_request(
-            "DELETE",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test:other.example.com",
-            access_token=self.access_token,
-        )
-        assert channel.code == 404, channel.result
-
-        # Test for a reasonable return value
-        channel = self.make_request(
-            "PUT",
-            "/_synapse/client/com.famedly/tim/v1/contacts",
-            {
-                "displayName": "Test User",
-                "mxid": "@test:other.example.com",
-                "inviteSettings": {
-                    "start": 0,
-                },
-            },
-            access_token=self.access_token,
-        )
-        assert channel.code == 200, channel.result
-        assert "displayName" in channel.json_body, "Result contains contact"
-
-        # 200 after adding
-        channel = self.make_request(
-            "DELETE",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test:other.example.com",
-            access_token=self.access_token,
-        )
-        assert channel.code == 204, channel.result
-
-        # 404 on unknown contact
-        channel = self.make_request(
-            "DELETE",
-            "/_synapse/client/com.famedly/tim/v1/contacts/@test:other.example.com",
-            access_token=self.access_token,
-        )
-        assert channel.code == 404, channel.result
-
-
-class FederationDomainSchemaTest(ModuleApiTestCase):
-    """
-    Test that the required fields for the federation list are present and parsable.
-    See:
-    https://github.com/gematik/api-vzd/blob/main/src/schema/FederationList.json
-    for the schema to use.
-
-    As of the time of this writing, these are fields that are required:
-        domain: str
-        telematikID: str
-        timAnbieter: str
-        isInsurance: bool
-
-    """
-
-    def prepare(
-        self, reactor: MemoryReactor, clock: Clock, homeserver: HomeServer
-    ) -> None:
-        super().prepare(reactor, clock, homeserver)
-
-        self.invchecker = InviteChecker(
-            self.hs.config.modules.loaded_modules[0][1], self.hs.get_module_api()
-        )
-
-    async def extract_entry_from_domainList(self, domain: str) -> FederationDomain:
-        """
-        Search for a specific domain in the federation list
-        """
-        _, raw_fed_list = await self.invchecker.fetch_federation_allow_list()
-        assert len(raw_fed_list.domainList) > 0
-
-        for domain_entry in raw_fed_list.domainList:
-            if domain_entry.domain == domain:
-                return domain_entry
-
-        assert False, f"Not found in federation list {domain}"
-
-    async def test_federation_list(self) -> None:
-        """Ensure we can properly fetch the federation list"""
-
-        domains, _ = await self.invchecker.fetch_federation_allow_list()
-        assert "timo.staging.famedly.de" in domains
-
-    async def test_common_fed_domain(self):
-        # First test the most common FederationDomain entry
-        # {
-        #     "domain": "timo.staging.famedly.de",
-        #     "telematikID": "1-SMC-B-Testkarte--883110000147435",
-        #     "timAnbieter": "ORG-0217:BT-0158",
-        #     "isInsurance": false
-        # },
-        test_entry = await self.extract_entry_from_domainList("timo.staging.famedly.de")
-        assert test_entry.domain == "timo.staging.famedly.de"
-        assert test_entry.telematikID == "1-SMC-B-Testkarte--883110000147435"
-        assert test_entry.timAnbieter == "ORG-0217:BT-0158"
-        assert test_entry.isInsurance is False
-
-    async def test_insurance_fed_domain(self):
-        # Then test an insurance FederationDomain entry. Want isInsurance to be True
-        # {
-        #     "domain": "ti-messengertest.dev.ccs.gematik.solutions",
-        #     "telematikID": "5-2-KHAUS-Kornfeld01",
-        #     "timAnbieter": "ORG-0001:BT-0001",
-        #     "isInsurance": true
-        # },
-
-        test_entry = await self.extract_entry_from_domainList(
-            "ti-messengertest.dev.ccs.gematik.solutions"
-        )
-        assert test_entry.domain == "ti-messengertest.dev.ccs.gematik.solutions"
-        assert test_entry.telematikID == "5-2-KHAUS-Kornfeld01"
-        assert test_entry.timAnbieter == "ORG-0001:BT-0001"
-        assert test_entry.isInsurance is True
-
-    async def test_illegal_fed_domain(self):
-        # This test is against a FederationDomain entry with data that is counter to
-        # what the schema says. In this case, 'timAnbieter' should be required but is
-        # reflected as `null`
-        # {
-        #     "domain": "messenger.spilikin.dev",
-        #     "telematikID": "1-SMC-B-Testkarte-883110000096089",
-        #     "timAnbieter": null,
-        #     "isInsurance": false
-        # },
-
-        test_entry = await self.extract_entry_from_domainList("messenger.spilikin.dev")
-        assert test_entry.domain == "messenger.spilikin.dev"
-        assert test_entry.telematikID == "1-SMC-B-Testkarte-883110000096089"
-        assert test_entry.timAnbieter is None
-        assert test_entry.isInsurance is False
-
-
-class MessengerInfoTestCase(ModuleApiTestCase):
-    async def test_default_operator_contact_info_resource(self) -> None:
-        """Tests that the messenger operator contact info resource is accessible"""
-
-        channel = self.make_request(
-            method="GET",
-            path="/_synapse/client/com.famedly/tim/tim-information",
-        )
-
-        assert channel.code == 200, channel.result
-        assert channel.json_body["title"] == "Invite Checker module by Famedly"
-        assert channel.json_body["description"] == "Invite Checker module by Famedly"
-        assert channel.json_body["contact"] == "info@famedly.com"
-        assert channel.json_body["version"], "Version returned"
-
-    @synapsetest.override_config(
-        {
-            "modules": [
-                {
-                    "module": "synapse_invite_checker.InviteChecker",
-                    "config": {
-                        "title": "abc",
-                        "description": "def",
-                        "contact": "ghi",
-                        "federation_list_url": "https://localhost:8080",
-                        "federation_localization_url": "https://localhost:8000/localization",
-                        "federation_list_client_cert": "tests/certs/client.pem",
-                        "gematik_ca_baseurl": "https://download-ref.tsl.ti-dienste.de/",
-                    },
-                }
-            ]
-        }
-    )
-    def test_custom_operator_contact_info_resource(self) -> None:
-        """Tests that the registered info resource is accessible and has the configured values"""
-
-        channel = self.make_request(
-            method="GET",
-            path="/_synapse/client/com.famedly/tim/tim-information",
-        )
-
-        assert channel.code == 200, channel.result
-        assert channel.json_body["title"] == "abc"
-        assert channel.json_body["description"] == "def"
-        assert channel.json_body["contact"] == "ghi"
-        assert channel.json_body["version"], "Version returned"
